+++ thunderstore/CHANGELOG.md
@@ -1,6 +1,7 @@
 Hint: this mod is still in Alpha - use at your own risk! These updates can break things that worked before. Still needs extensive testing. In case of any errors please create an issue: https://github.com/Kandru/mimesis-player-enhancements/issues
 
 ## 26.7.22
+- fix: properly show net-worth of spectating player
 - draft: add monster spawner via the dashboard
 - draft: when waiting for all players to load into the dungeon everybody could already move. Now they're freezed properly.